ai-services/pill-identification/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
google-cloud-vision==3.4.4
Pillow==10.1.0
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
scikit-learn==1.3.2
tensorflow==2.15.0
